Add Compatibility to Python3 for metaclasses
Python3 changed the way how to use a metaclass:

See
https://python-future.org/compatible_idioms.html#metaclasses

On python2, one can test it on classes using the 'InputSingleton':

MythTV is using metaclasses for getting the time zone info
in utility/dt.py: The `class posixtzinfo` is designed as
`InputSingleton`, which means that every subsequent call
with the same parameter returns the same object.
With the patches applied from #13299, one gets on python2:

    $ python
    Python 2.7.15+ (default, Nov 27 2018, 23:36:35)
    >>> from MythTV.utility.dt import *
    >>> from MythTV.utility.singleton import InputSingleton

    >>> a = posixtzinfo()
    >>> b=  posixtzinfo('Etc/UTC')
    >>> c = posixtzinfo("America/Anchorage")

    >>> x = posixtzinfo()
    >>> y = posixtzinfo('Etc/UTC')
    >>> z = posixtzinfo("America/Anchorage")

    >>> c == z
    True
    >>> a == x
    True
    >>> b == y
    True

With python3, i get:

    >>> a == x
    False
    >>> b == y
    False
    >>> c == z
    False

which means, the 'InputSingleton' is not applied in the correct way.

Python3 simply ignores the '__metaclass__' property of Python2.

This patch adds the compatibility layer from python-future.

Additionally, this patch fixes an error

  'RuntimeError: dictionary changed size during iteration'

which is needed for python2 to python3 conversion and only occurs if
the '__metaclass__' is not applied correctly (on python3).

(cherry picked from commit d6af02f8e93aaad26e6289a6634a8ab1f033117e)